refactor(predict): relocate PredictPortfolio module into PredictHome view
Move the reusable PRED-835 PredictPortfolio module from
components/PredictPortfolio into views/PredictHome/components/PredictPortfolio,
since the redesigned homepage is its only consumer. Pure relocation: updates
the moved files' relative import depth (now two levels deeper) with no behavior
change. PredictHome still renders the existing mock placeholder; wiring follows
in a later commit.
