+++ app/components/UI/Predict/views/PredictHome/components/PredictPortfolio/PredictPortfolioModule.tsx
