+++ app/components/UI/Predict/views/PredictHome/components/PredictPortfolio/index.ts
